Wrote basic ball tracking code still some bugs
So we are able to get ball tracking to work for the 2 cases
we want it to work for but we are getting some bugs that we aren't
sure about. Jump squat is kinda difficult to manage for side
tossses where it sometimes doesn't recognize that a side toss
happens and never leaves jumpsquat.
We made the circles on the video be based off of the CircleObjects
instead of the blobs and now they tell us if the state is
airborne(green) or jumpsquat(red)